langgraph/pyproject.toml
[tool.poetry]
name = "langgraph-loganalyser"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
black = "^24.4.2"
isort = "^5.13.2"
langchain = "^0.2.6"
langgraph = "^0.1.5"
langchain-google-vertexai = "^1.0.6"
langchain-community = "^0.2.6"
groq = "^0.9.0"
langchain-google-community = "^1.0.6"
langchain-groq = "^0.1.6"
pyowm = "^3.3.0"
tavily-python = "^0.3.3"
langchain-experimental = "^0.0.62"
google-cloud-logging = "^3.10.0"
pandas = "^2.2.2"
pytz = "^2024.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
